v4/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.1.0
langgraph==0.0.20
google-generativeai==0.3.2
google-cloud-vision==3.5.0
pgvector==0.2.4
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.1.0
rank-bm25==0.2.2
numpy==1.26.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
flake8==7.0.0
pre-commit==3.6.0
gunicorn==21.2.0
django-cors-headers==4.3.1
